Questions/deploy/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[argon2]
python-dotenv
pymysql
pycryptodomex>=3.20.0  